Update patch set 1
Patch Set 1:

Dirk and I are trying to get 100% #COMPONENT coverage under chromium/src and identified third_party/libsrtp as a folder without a # COMPONENT entry in it's OWNERS file. 

Can you take a look at this and make the changes in your repo? I unfortunately do not have access to commit this change.

Thank you for your help.

Patch-set: 1
Work-in-progress: false